refactor abstract adapter and reader, add block reader
